requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==5.0
django-jazzmin==2.6.0
django-paypal==2.0
django-phonenumber-field==7.3.0
idna==3.6
paypalrestsdk==1.13.3
phonenumbers==8.13.27
Pillow==10.1.0
pycparser==2.21
pyOpenSSL==23.3.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
